backend/requirements.txt
# Web
fastapi
uvicorn[standard]

# Data
numpy
pandas
pyarrow

# ML / audio
torch
torchaudio
librosa
scipy
scikit-learn
laion-clap
umap-learn
